r1228: use int64_t instead of long long
(jra: please use: void, char int, uint_t, [u]int<8|16|32|64>_t types in new code)

metze

+++ source/torture/basic/denytest.c
@@ -1464,7 +1464,7 @@ BOOL torture_denytest1(int dummy)
 		}
 
 		if (torture_showall || res != denytable1[i].result) {
-			long long tdif;
+			int64_t tdif;
 			GetTimeOfDay(&tv);
 			tdif = usec_time_diff(&tv, &tv_start);
 			tdif /= 1000;
@@ -1558,7 +1558,7 @@ BOOL torture_denytest2(int dummy)
 		}
 
 		if (torture_showall || res != denytable2[i].result) {
-			long long tdif;
+			int64_t tdif;
 			GetTimeOfDay(&tv);
 			tdif = usec_time_diff(&tv, &tv_start);
 			tdif /= 1000;
